+++ packages/spacecat-shared-tier-client/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@adobe/spacecat-shared-tier-client-v1.3.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-tier-client-v1.3.0...@adobe/spacecat-shared-tier-client-v1.3.1) (2025-11-24)
+
+
+### Bug Fixes
+
+* error message on org not found ([#1165](https://github.com/adobe/spacecat-shared/issues/1165)) ([c1a7702](https://github.com/adobe/spacecat-shared/commit/c1a77026a485b4ce483b9ce806bcf6e722064807))
+
+# [@adobe/spacecat-shared-tier-client-v1.3.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-tier-client-v1.2.5...@adobe/spacecat-shared-tier-client-v1.3.0) (2025-11-21)
+
+
+### Features
+
+* added async export for tier-client ([#1136](https://github.com/adobe/spacecat-shared/issues/1136)) ([4098f39](https://github.com/adobe/spacecat-shared/commit/4098f394cb526d514e502d9c389168480a0987df))
+
 # [@adobe/spacecat-shared-tier-client-v1.2.5](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-tier-client-v1.2.4...@adobe/spacecat-shared-tier-client-v1.2.5) (2025-11-18)